Replaced the old project with a new one
I replaced the project with a new one thats not on the universal render pipeline because i had a certain error because of a missing dll file. I tried reinstalling unity but it didnt work. I also tried getting the dll file myself but i couldnt figure out exactly what file was missing and where i need to install it to. There was also nothing on the internet about this error so i just created a new project on the regular 3d pipeline and moved the assets folder over and the problem was fixed.